Zuul: [mediawiki/extensions/Wikibase] Remove stray experimental job
This was added way back in change Icb87b9d429 (commit 771b67bc8e), when,
I assume, the rest of CI was still on some older PHP version (the other
job names don’t seem to include a PHP version so it’s a bit hard to tell
what it was and I can’t be bothered to figure it out). These days this
is one of the standard test jobs and there’s no reason to have it in
experimental too.

Change-Id: I8e9e64b043e70eb952d6344b667e8a34040cfeb9