dockerfiles: rm buster-backports and use component/shellcheck
The Debian project has removed `buster-backports` which we used to
install the shellcheck utility. We had it imported in our
`buster-wikimedia` distribution under `component/shellcheck`.

Add our apt component and the shellcheck pinning in our base `ci-buster`
image which share the config for `operations/puppet` and `tox-buster`.

This also unbreak the rebuilding of Buster based image which failed
with:

  E: The repository 'http://mirrors.wikimedia.org/debian buster-backports
  Release' does not have a Release file.

Bug: T362518
Change-Id: Ied00bc6d023bb26cabcf8fc789815b1da555b77b