zuul: [mediawiki/extensions/CampaignEvents] Add WikimediaCampaignEvents dependency
Run WikimediaCampaignEvents tests as part of the CampaignEvents build,
so that we can be more confident that changes to the main extension do
not break the WMF-specific functionality.

Note that the base extension is, strictly speaking, not dependent on the
WMF-specific one (in fact, it's the opposite, and this is already
reflected further down in the dependency list).

Change-Id: I77607709c6b66f2b2d47a849898ed09826b06a2f