mwcore-codehealth: ensure cover directory exists
In I599da09ee27aade02ce8c43facf6535fcad7b280 we wanted to ensure that
directories exist before they're mounted as volumes by docker because
docker will create directories that don't exist if you mount them as a
volume, resulting in unexpected permission problems.

The `mwcore-codehealth-master-non-voting` job mounts a `cover`
directory, but does not create it first. Adding `docker-wipe-dir`
ensures that a directory is created with the proper permissions and
empty before mounting as a volume.

Affected jobs:
 - mwcore-codehealth-master-non-voting
 - mwcore-codehealth-patch

Bug: T365260
Change-Id: I82714a120f5f78ab64a3e3e4fa36cd709e97f63d