Enable parallel phpunit for WikibaseLexeme
Quibble 1.9.0 introduces support for running some PhpUnit test
suites in parallel by setting the QUIBBLE_PHPUNIT_PARALLEL
environment variable. Since running many jobs in parallel places
significant extra load on the CI infrastructure, we don't want to
just roll this out to all jobs simultaneously.

Here, we selectively set the environment variable just for the
`WikibaseLexeme` project - a repo that gets relatively few
patches. The team maintaining Lexeme is aware of the coming CI
changes and can support with the evaluation of the impact of the
change. If the initial rollout is successful, this can be extended
to other repositories.

Bug: T361190
Change-Id: I331790c0b43d73972200d8a41b6538e17a463cd9