dockerfiles: castor: always restore to /cache
With Nodepool, we had to vary where the cache was restored and relied on
the job name having `docker` to find when we ran in a container. With
the removal of the `-docker` suffix from job names (T360327) that caused
the rsync to restore to $HOME which is not existent and effectively
DISABLED RESTORATION OF CACHE!

Always restore to `/cache`.

While at it, update CASTOR_HOST to the current instance hostname.

Bug: T368550
Change-Id: I4955cdfa7cff572c808795f9ff494ea3b792906a