Add Wikibase client-only jobs to EntitySchema checks
EntitySchema can be installed as an extension on Wikis running
Wikibase in client-only mode. We need to test that nothing breaks
when the extension is installed on client wikis, and that the
client wikis can then resolve and process `EntitySchema` objects.

Add the existing Wikibase client-only job to the EntitySchema
checks, and add `EntitySchemaClient` to the list of test groups
to run. This will have no effect for the existing job on Wikibase
checks (since no tests with the group exist), but should allow
us to tag PHPUnit tests that we want to include in the check with
`EntitySchemaClient` in the `EntitySchema` repo to have them run.

Affected jobs:
 - wikibase-client-php74
 - wikibase-client-php81
 - wikibase-client-php83

Bug: T367156
Change-Id: I93a08d9e8c88aacd2ec4dce8a052343fbf365a13