Move wmf/branch_cut_pretest in gate-and-submit-wmf
The test-wmf and gate-and-submit-wmf have a strict branch filter which
is not matching the "newish" `wmf/branch_cut_pretest`. That branch is
cut on a daily basis and aims at preparing for production, we should
thus launch the same set of jobs we do for the main wmf/* branches.

Adjust the branch filters to match `wmf/branch_cut_pretest`.  This drops
jobs using a php version we do not use in production.

Change-Id: I8393c5027805721b1c0c189142162f6cfa82348a