castor-save-workspacecache: check remote dir exists
Since we use docker to run rsync on the remote, currently, if a cache
directory has already been removed, then castor-save-workspacecache will
create the directory as root (by running docker --volume).

This mirrors what we do in `docker-wipe-dir` in
`jjb/macro-docker.yaml`—before mounting a volume with docker, check if
it exists. If the directory is absent, do not mount it to a docker
container since it will then be created by root.

Bug: T282893
Change-Id: If4d1dffefbdda30491a8f9716fd4ffdc5d8b819f