jjb: remove unused castor-save publisher
The `castor-save` publisher was to trigger a job that got removed back
in July 2020 with ea1b1859efa5ee.  That comes from a time we ran
Nodepool and all that got replaced by a variant that requires $WORKSPACE
to be set (`castor-save-workspace-cache`).

Adjust a comment that got copy pasted from the Nodepool era, the save
has to be put on hold before deleting the cache files.

Change-Id: Id758c39efe8c97d1527d45e80a31d57938ce2760