jjb: change mw coverage publish jobs to use setup/teardown
Use the setup to macro to create the directories and load cache. This
slightly reorder the commands used that does not matter.

Same for teardown which also save the cache when necessary.

That resolves a FIXME from 8defb898794f24f which at the time did not fix
everything in order to be a noop.

Affected jobs:
 - mwext-phpunit-coverage-publish
 - mwskin-phpunit-coverage-publish

Change-Id: I4a7c0d5c9af250922545ccad3bd22d3e799268ed