jjb: grep from log dir to have short names
In 44a3dc6384e the files are found using a glob pattern and I suggested
to prefix them with "$WORKSPACE"/log which was to also save us from
having to change directory.  The error files are thus absolute paths and
that is quite long in the console output.

Change to $WORKSPACE/log when globbing the error files and when doing
the grep. It will then show the file names relatively to the log dir.

Change-Id: Ic0faa4a95cc43ce14bef3c2ed37599e884d54ba5