Add xdebug back to Quibble PHP images
We previously removed xdebug from all images since there was no
use case for enabling the extension in the CI images.

For Quibble images, it can be helpful to attach a debugger to a
local run of a CI job, to more exactly reproduce test failures.

Add xdebug back to the Quibble CI images.

Bug: T319495
Change-Id: Ib2b2171e0c43c11358f9bea80b53ebc521c49eea