Automoderator: Make CommunityConfiguration a dependency
Automoderator uses CommunityConfiguration for many of its configuration
needs in Production, and so it makes sense for them to be tested jointly
as well. In addition, we want to add a test-file that asserts that
Automoderator's configuration schema is plausible and that it is in sync
with its pre-calculated emergency backups (see the change
Ic99b6ab0e9f7717fa286822aba76b9afe456f1a1).

Bug: T378092
Bug: T374428
Bug: T367516
Change-Id: I2e2ac0b3fc18ee0296b49e2a225e8f631ce84a12