zuul: fix finding Zuul config dir
a541299aae3c introduced YAML files to express dependencies and in order
to load the file it needs the path to the directory holding the file.
Since Zuul read the Python file, compile it and then pass it without any
global set, the `__file__` variable is not set. In the interest of time,
the code was written with a brute force approach with the following
warning mention:

  HACK: This is horrible, but __file__ isn't defined.

I had the use case of loading the file locally:

  from parameter_functions import dependencies, get_dependencies
  print("\n".join(sorted(
    get_dependencies('GrowthExperiments', dependencies)
  )))

Which aborts with:

  ZUUL_DIR could not be set, environment not recognised.

Zuul compiles the code using `compile()` which requires a filename to be
given, the information is thus hidden in the compiled bytecode.
Although not directly available via __file__, that can be retrieved by
inpsecting the current frame using:

  inspect.getfile(inspect.currentframe())

Replace the logic with the above one liner.

Reorder the imports, splitting the non builtint "yaml" on its own line.

ZUUL_DIR sounds like the directory of the Zuul application, rename it to
ZUUL_CONFIG_DIR to disambiguate.

Change-Id: I494f4b0a2ca1096c2a4078023e05a98370224904