Enable PHPUnit Parallel tests for WikiLambda
WikiLambda has been excluded from PHPUnit parallel testing because
of test failures that we were seeing when the feature was
introduced. Since then, we have made multiple changes to the way
the test suites are split and ordered that make inter-extension
test issues less likely to occur.

In limited testing, the parallel tests now appear to pass for
WikiLambda. Remove the exception that disabled parallel testing.

Bug: T375551
Change-Id: I2176f640c326cd3cdda1de0f37b905c78a234ea1