Docker: [operations-puppet] Fix mv logspam
When we run multiple tox modules (such as when you edit the Rakefile
or taskgen.rb), the mv command in the run.sh script will fail because
it tries to move many different 1-install_deps.log etc files to the
same destination directory.

This doesn't actually cause the build to fail, but does make the
output very noisy, and hides actual failures behind a lengthy
red herring.

Change-Id: I6dace75fd39b6f8592ac5667545eb0e2997fefcd