Zuul: allow archival of BlueSpice repositories
fc444e2078 introduced a test to verify BlueSpice repositories only use
the bluespice template. It introduced a hardcoded list of archived
repository which thus had to be updated whenever we archived a repo.

Check whether the template is 'archived' to detect a repository has been
archived and remove the hardcoded list from the test.

Change-Id: I32a3945199274071d9348efc4ae301d2bdfd649b