zuul: Improve "PHP test coverage" comment by jenkins-bot
For the success-message, use `OK!` instead of `increased (…) :-)`
so that it is easy to spot without as much cognitive processing and
reading each word. In particular, the emoji at the end seemed to get
lost to on me quite often given the visually greyed-out styling
and other nearby parentheses, dashes, and colons.

Example

> Timo          Uploaded patch set 1.
> jenkins-bot   Added to cc
> jenkins-bot   Performance checks OK! (same or better) - https:/integr...
> jenkins-bot   PHP test coverage increased (or stayed the same) :-) - https:/.
> SonarQube Bot Added to reviewer

Also matches the perf checks this way.

For the failure-message message, clarify that this can also be a
build failure. There is a fairly high chance of this being the case
because patch coverage is one of the fastest jobs to complete, and
thus often beats other jobs reporting syntax/runtime failures.

This, coupled with the seemingly confident message that coverage
decreased has misinformed several patch authors into thinking coverage
has decreased and started looking for where it might be decreased
in their code (instead of cehcking the job output first whether the
build simply failed before it could generate a coverage report).

Change-Id: I5ea1489c00310f224bb0d148116c2b54b0d11fb6