jjb: remove quibble-integration Matrix job
I have created the job six years+ ago in order to quickly test Quibble
over different branches, databases backend and php version. Quibble has
long been a stable software and I no more have a use for this. That also
has the benefit of removing a Matrix job from the setup which I guess
are not trivially portable to a newer Zuul.

Inline the `mediawiki_supported_branches` YAML alias which was reused
for the php-security jobs.

Deleted jobs:
 - quibble-integration

Bug: T383134
Change-Id: I2a92085dab047b7b17872131c64841ca13a88c52