Remove debian-glue-nocheck job
The job passed DEB_BUILD_OPTIONS='nocheck' to the Debian packaging
toolchain in order to skip running tests. It is a bit of an anti pattern
but at the time I never managed to have the tests to pass (T222952). We
have since moved the Zuul deployment to scap (T215458) and thus no more
need debian-glue for it.

Bug: T215458
Bug: T222952
Bug: T383134
Change-Id: Ic5566fdfcbe1c55c5ae9cda98cb82ebc8bb065e5