Remove php-security scanning
As mentionned in T309655, those jobs results are not monitored by
anyone. The initiative dates from 2014 (T74193), was briefly worked on
to expand (T180278) but never was generalized nor became a process.

The Wikimedia security team is working on dashboard which scans
repositories (T371814).

Deleted jobs:
 - fundraising-php-security
 - github-php-security
 - mediawiki-php-security
 - php-composer-security
 - shellbox-php-security

Delete releng/composer-security-check image as a result.

Bug: T180278
Bug: T309655
Change-Id: I1d53f217d675a2ad6b78c4abc7935a6681ee568b