zuul-migrate: import script from upstream
The script was a one off script to migrate to Zuul v3 and hasn't been
further maintained, but that is still a base we can use. Note it
depended on JJB 1.6.2 when we use 4.3.0.

Script imported AS-IS [from Zuul]. A following commit will do some fixes
to make it runnable and adjust for the newer JJB.

Disable indent_size check in editorconfig since the style uses
left-padding.

[from Zuul]
https://opendev.org/zuul/zuul/commit/f1e7ffab934b7522111d1ec2e5e0e8ffe3c29edc

Bug: T383273
Change-Id: I84879068e32d6e83a2e0d2669f2d7006ee5e15d8