Zuul: [mediawiki/extensions/CirrusSearch] Drop WikibaseCirrusSearch dep
WBCS depends on CS, but this relationship is also inverted, so any jobs
that trigger CirrusSearch pull in a lot of dependencies unnnecessarily.

The gate will be unaffected, as Wikibase and its dependencies already
rely on CS, and after this change, on WBCS and CS through it rather
than the other way around, but regular jobs' dependency trees will now
be a little slimmer, which might help with the issues GrowthExperiments
have been having around test load.

Bug: T386015
Change-Id: I0cf4fb0078a951027e30a422f7b994528bf17645