Fix `suite.xml` coverage configuration generation for Wikibase
The `*-coverage` images use the `phpunit-suite-edit.py` script to
add coverage directories to the PHPUnit `suite.xml` configuration
file. By default, the folders `src`, `includes` and `maintenance`
underneath the extension root are added for coverage when generating
coverage reports for Mediawiki extensions.

In the case of Wikibase, none of those folders exist. Because all
of the folders then listed in `<coverage><include>` are missing,
`phpunit --coverage-*` reports "Incorrect filter configuration,
code coverage will not be processed". The coverage report is not
generated, a subsequent existence check for the report fails, and
the job errors.

Change `phpunit-suite-edit.py` to test for the existence of folders
before adding them to the `<coverage><include>` tag. If no folders
are added (because none are found), add the extension root to the
include so that we get some coverage information.

Bug: T288396
Change-Id: I40c99b4e45ce6cff78c05bbf542ee804a7a8ea48