Fix patch coverage generation for Wikibase
The `*-coverage` images use `phpunit-patch-coverage` to generate
coverage information for submitted gerrit patches. This script
assumes a default PHPUnit test location of `tests/phpunit` in the
project / extension root.

In the case of Wikibase, this folder does not exist, causing
`phpunit-patch-coverage` to fail.

Change `mwext-phpunit-coverage-patch.sh` to check to see if the
`tests/phpunit` folder is present, and in the case that it is not,
fall back to using the extension / project root.

See also I40c99b4e45ce6cff78c05bbf542ee804a7a8ea48

Bug: T288396
Change-Id: I07327632ddcb869b4b0d440c89c42883f3313362