Zuul: Start testing in PHP 8.4 for 'mediawiki-php-library' repos
There are three libraries that fail, each of which are left on
PHP 8.3 for now with a comment (one task specific to the repo,
and two repos with the same phan issue already filed).

Rename the generic template to specify the versions, like we did
for the composer templates when we added php84 variant jobs as
voting.

Bug: T386108
Change-Id: Ia86c90b1d66270ccdc0e892ad7fefa8fb5b60414