utils: script to manually launch jobs
The script runs PHPUnit and Selenium jobs for any repository having a
dependency set.  This is done by submitting requests to the Zuul Gearman
server, it report back the repo/job/success.

This is in order to test Ia268663f87efbf5f6d4e37a4a4ffff2efddbb16e
before it is deployed.

Bug: T389998
Change-Id: I2aa24920a1967e812627597978e97ed82c9afb34