Docker: [quibble-bullseye] Switch MariaDB to 10.6 Wikimedia package
We've been wanting to test CI against a more modern version of MariaDB for
a while. All of production is on 10.6 or 10.11 (in test, see T365805). The
default in Debian Bullseye is unfortunately 10.3, so we've been only using
this old version. Switching over to 'wmf-mariadb106' not 'mariadb-server',
as a drop-in replacement, lets us test this in CI (albeit some years late)
and drop 10.3 support requirement from MediaWiki as soon as we drop Buster
(i.e. PHP 7.4) testing, too.

Bug: T366646
Change-Id: Ic24c8fe1dec7a52376227595adf935dc07135ab5