Rename prod Gerrit connnection to match config
When configuring Pipeline connections, the names refer to the
connection names given in zuul.conf.

If the name of the connnection that received the event mismatches, the
event is rejected (T391374#10750390)

Since `gerrit` could be misread as "a connection with the driver
gerrit", I have renamed the connections in zuul.conf to have meaningful
names:

  [connection "gerrit-dev"]
  server=gerrit
  ...
  [connection "gerrit.wikimedia.org"]
  server=gerrit.wikimedia.org
  ...

This is the counterpart renaming the connection in the pipelines
configuration.

Change-Id: I27395507ae070f84fe57e3a45c7bdb3381a7ad7b