Replace all uses of `$(pwd)` with `$PWD`
Under Bash, the `pwd` builtin just prints the value of the `$PWD`
variable, so `$(pwd)` is exactly the same as `$PWD`. Under other
shells they are only *mostly* the same, but we probably don't
use any other shells, and it'd probably be close enough anyway.
Furthermore, `$PWD` avoids the overhead of launching a subshell,
and it avoids output when using `set -x` to print commands before
executing them. This last benefit is my reason for this change.

https://unix.stackexchange.com/a/175000
https://unix.stackexchange.com/a/453268

Bug: T393847
Change-Id: I0d8eea338c21d4d74bbccd607bb079b80daee784