dockerfiles: silence git
Silence git checkout when building the php Ast extension. This only
affect the build output so the image is not being rebuilt via a
changelog entry.

While at it, expand git checkout `-q` option to `--quiet` in the
quibble-fresnel script. That is to be consistent with the rest of the
repository.

Bug: T393847
Change-Id: Ifb0ddd3cce2b8d08786d64572b5a0f006f483270