dockerfiles: silence git in ci-src-setup
`git init` emits a warning about the initial branch name. Set it
explicitly to `$ZUUL_BRANCH` with a fallback to `master` (since updates
for tags might not have it set).

Only update images actually using `/utils/ci-src-setup`.

While at it, expand git checkout `-q` option to `--quiet` to be
consistent with the rest of the repository.

Bug: T393847
Change-Id: I401a2866494aa4e0d5e8d127d8bc6fd45c1769dd