Zuul: Upgrade all Quibble 'apitests' jobs from 7.4 to 8.1
The only bits left in production are for maintenance script, the API has
been running with PHP 8.1 for sometime now and we thus no more need to
guarantee they work with 7.4.

This also replaces the base image from Buster to Bullseye, and so to an
image that has OpenSearch (needed for T386691).

Bug: T386691
Change-Id: I3533fea7ea0c78537a1e873e86345eb9e2f4a96b