jjb: Skip first notif for Selenium jobs on beta
Extend the `selenium-daily-{sitename}-{project}` template to take four
variables controlling when emails are sent:
* failure: send email on each failure
  (default: true)
* still_failing: send email on 2 or more failures in a row
  (default: false)
* unstable: send email when unstable (unit test failures)
  (default: true)
* still_unstable: send email when unstable for 2 or more builds in a row
  (default: false)

The `selenium-daily-beta-MediaWiki` and
`selenium-daily-betacommons-MediaWiki` tests are now configured to use:
* failure: false
* unstable: false
* still_failing: true
* still_unstable: true

The changes to the `selenium-daily-beta*-MediaWiki` jobs will stop
emailing folks when the job flips from success to failure for the first
time. Instead notification will require multiple consecutive failures.
In T270771#10829124 Željko explained that these tests are only
interesting to him when they fail at least two days in a row. Changing
the notifications to match that expectation is hoped to reduce
notification fatigue for developers and testers.

Affected jobs:
* selenium-daily-beta-MediaWiki
* selenium-daily-betacommons-MediaWiki

Bug: T394551
Change-Id: I8c9922e6c6f86e18fb87fb29e905126d5af3e613