jjb: set cache directory ownership to nobody:nogroup
When Castor restores the cache from the central cache on the Jenkins
agents, it spurts a permission issue when setting times for /cache/..:

  Syncing...
  rsync: failed to set times on "/cache/.": Operation not permitted (1)
  rsync error: some files/attrs were not transferred (see previous
  errors) (code 23) at main.c(1668) [generator=3.2.7]
  Done

The reason is the cache directory is created by the Jenkins agent
running as `jenkins-deploy` but Castor rsync in a container with the
`nobody` user. The directory should always have matched the uid/group of
the container.

Change the `mkdir` running as jenkins-deploy to a container invocation.

This solves T188488: rsync failing to set times on "/cache/."

Would partially solve T282893 which is the directories being unwritable
due to them being owned by root (due to some race condition).

Bug: T188488
Bug: T282893
Change-Id: Ide99ef472afd8d1abe097de735071f621d3fe024