jjb: gerrit: use standard "setup" macro
For Gerrit, we need to change the remote URL when processing submodules.
The extra parameter given to `docker-ci-src-setup-simple` prevented to
switch to `setup`/`teardown`.

Make the `setup` macro to pass `submodules` down to
`docker-ci-src-setup-simple` and simplify the `gerrit` job as a result.

No job is affected.

Follows-Up: I5236933e5bcb6690e3073409352d8b4441173860

Change-Id: I6e64e47def78cab17abba87b7720d6427bd92f56