Docker: [node-test-brower-php*-composer] Build php-ast inline
Due to an issue in the upstream PHP release process, the values of
tokens are changed between patchset versions of PHP 8.3. The reason is
the releases were made by different persons and one of them had a way
older version of GNU Bison which results in different values being
generated for the tokenizer tokens.

Upstream is aware of the issue:
  https://github.com/php/php-src/issues/14937

The token values in AST are set at compile time and borrowed from the
PHP version used to compile it. Thus if we rebuild php-ast and install
PHP at the same time, we should have matching token values.

Add a PHP script to verify the token `T_CLASS_C` from PHP and
`\ast\flags\MAGIC_CLASS` from the ast extension do match. The script is
added in a scratch image `php-scratch` similarly to what we have done
with `composer-scratch`. The PHP images depend on it to ease
rebuilding.

Example output for the test script:

    $ php dockerfiles/php/test-ast-tokens.php
    Verifying AST extension matches PHP...
    See: https://phabricator.wikimedia.org/T396312#10902917

    Reading /usr/lib/php/20220829/ast.so ... ok

    | Versions     | Constant               | Value
    |--------------|------------------------|------
    | PHP 8.2.28   | T_CLASS_C              | 346
    | AST 1.1.0    | \ast\flags\MAGIC_CLASS | 346

    Verifying whether T_CLASS_C === \ast\flags\MAGIC_CLASS ... ok

Follows-up-to: If0abfacaf54cd635da2018a9bb886e861159819f

Bug: T396312
Change-Id: I5624f24f79de39991c7f15b0f3a58e7f66fcd3f0