dockerfiles: Re-disable pcov in php81-84 base images
The incompat with xdebug (whch we use in a few places) makes this
harder than needed. If we want to solve T395743 with built-in
phpunit validation for `@covers`, despite upstream PHPUnit declining
this use case, I guess we'll need to go the way of enabling it in
the composer-test* images. Or we can give the Phan/PHPCS route a try.

Bug: T395743
Change-Id: Iaa5e5b6e0562135dbdda2ed336d9b6aa665aacb0