jjb: move publish-to-doc fetching to a macro
This move the logic to retrieve files from a remote agent to a macro
"fetch-from-remote-instance". The intent is to reuse the logic to
retrieve Prometheus metric files and publish them.

The DOC_URL variable is only used for display purpose and is moved
"below" to the publish-to-doc job.

Affected jobs:
 - publish-to-doc

Bug: T392894
Change-Id: I008c9c76b462f97c479187a1e8434d5187e3e26b