jjb: doc-publish support retrieving specific files
When publishing documentation, we retrieve anything under $WMF_CI_SRC.
For retrieving Prometheus metric files, we are only interested in *.prom
files.

Introduce a new variable $WMF_FILES_PATTERN which is passed to rsync. It
default to `.` to retrieve the whole directory. That ends up being a
noop for `publish-to-doc` job.

Affected jobs:
 - publish-to-doc

Bug: T392894
Change-Id: I1d8ef1cd48276dcb6b2be0abd8a8e9588b388f24