jjb: allow missing source when fetching from instances
This change `fetch-from-remote-instance` to ignore missing sources when
doing a rsync. This is specifically made for the retrieval of Prometheus
files (T392894) which might not yet be generated by repos (T398284).

When no pattern is given, the source is `.` in which case we keep the
existing logic: we certainly always want the remote directory to exist.

Affected jobs:
 - publish-to-doc
 - publish-to-prometheus

Bug: T392894
Bug: T398284
Change-Id: If5bc145b4196ce4c109e217175b34b585b4a85c8