dockerfiles: rake-ruby2.7: add Vagrant dependencies to rake-ruby2.7
For mediawiki/vagrant, we have maintained a custom child image to add
zlib1g-dev (to compile Nokogiri) and rsync (needed by VagrantFile).

With the Zuul upgrade I want to reduce the number of images (T397429).

Add Vagrant dependencies in the generic `rake-ruby2.7` image. We can
then migrate the job from the custom `rake-vagrant-ruby2.5`.

Bug: T398469
Change-Id: I99f10b1e18f95b6f5ff155e3979771305dc59332