jjb: beta: remove obsolete tmp dir setup
The script does not seem to rely on having a specific temporary
directory. It came as part of replacing the old global scripts to setup
and teardown the workspaces: 3b25c6f22fc469e22

This simplify the job and also drop the last usage of EXECUTOR_NUMBER
environment variable in our system.

Affected job:
- beta-update-databases-eqiad

Bug: T399283
Change-Id: I152bd75c0d4ddee92f8a9ec27df4e2e2543bcd5f