jjb: remove GitHub API token from all jobs
Due to GitHub rate limiting unauthenticated users, we went to generate a
token attached to an unprivilege user.

The token is provided by the credentials plugin and exposed in jobs as
`COMPOSER_GITHUB_OAUTHTOKEN` environment variable. The images entry
points then should invoke `composer config` to set the OAuth
configuration (e0982e0c4).

There are some issues though:
* some images do not invoke composer config, notably the Quibble images
* if one uses a different entry point, they need to remember to set the
  credential

Composer authentication can be configured via `COMPOSER_AUTH`
environment variable which should be set to a json payload such as:

  COMPOSER_AUTH={"github-oauth":{"github.com":"<TOKEN>"}}

This can be set as a global environment variable and it will be exposed
to all jobs/containers.

Bug: T399846
Change-Id: I1b8963f768a9c2fd1aeecbc7bb1fc57ac5fa9570