jjb: Create specific job for Wikilambda daily tests
The wikilambda-catalyst-end-to-end job is triggered by Zuul and emits
emails.  With ef2536e098, the job was made to be triggered by Jenkins on
a daily basis.

The job emits email notifications, and it is hard to find whether a
failure came from an ongoing patch in Gerrit or whether it is an issue
with the Catalyst infrastructure.

Create an alias job wikilambda-catalyst-end-to-end-daily which runs on
a daily basis.

Add CATALYST_ENV_NAME parameter which is recognized by the catalyst
image 1.3.0 (see 8ba32ff).

New job:
 - wikilambda-catalyst-end-to-end-daily

Bug: T388056
Change-Id: I737d66ed761394874a42101fc70b185dcbc4262b