Remove composer timeout on wikibase phpunit jobs
The `wikibase-repo-php81` jobs have started regularly hitting the
composer timeout when running their PHPUnit tests (see details in
ticket). This could simply be that there are now more tests, or it
could be a sign that the CI system is overloaded at peak times -
on the same day we see the tests timeout at 5 minutes, we also see
runs of the test suite complete in 3.5 minutes.

The Wikibase jobs were switched to use `composer phpunit:entrypoint`
with b4c35da85 and PHPUnit is thus subject to composer 600 seconds
timeout.

Set the composer timeout for the command to zero so that it runs
to completion.

Affected jobs:
 - wikibase-client-php81
 - wikibase-client-php83
 - wikibase-repo-php81
 - wikibase-repo-php83

Bug: T401888
Change-Id: I9877ad78248fb8aebb195ac222ccd24e63791596