jjb: remove assert-lock experimental jobs
The job is intended to run composer install inside mediawiki/vendor and
assert there is no difference in the resulting installation . That was
made three years ago (T74952) and never went in use.

In the interest of reducing the number of jobs to migrate to the new
Zuul (T383134), remove it.

Deleted jobs:
 - composer-php81-assert-lock
 - composer-php82-assert-lock
 - composer-php83-assert-lock
 - composer-php84-assert-lock

Bug: T74952
Bug: T383134
Change-Id: I082995ea3cc8250b20c8b87974e1e1cda2ba2e03