quibble-coverage: make phpunit-suite-edit skin agnostic
phpunit-suite-edit is passed the extension or skin name. The code is
mostly duplicated, beside the extension vs skin path name, there is an
extra conditional for extensions. That was done in f4c0bd68a to handle
Wikibase (T288396). I guess it does not hurt for skins.

Remove --cover-skin
Make --cover-extension to be given the extension with the extension/ or
skin/ prefixes.
Update calling scripts

Bug: T402398
Change-Id: If96e432eca092533522175807e8fd58d5d6eedfb