quibble-coverage: pass --test-dir as a relative path
While trying to use `mwext-phpunit-coverage-patch` for skin, I found a
bug due to `--test-dir`. It has been introduced to fix coverage for
Wikibase (01565b366f8d for T288396) and went to use an absolute path.

When running the mwext variant:
```
$ phpunit-patch-coverage check \
  ... \
  --test-dir /workspace/src/skins/Vector/tests/phpunit
Finding coverage difference in xxx
Could not find any tests to run.
```

The reason is `mediawiki/phpunit-patch-coverage` looks up for file
changed in HEAD which are given as relative paths. They are filtered so
that they must start with the value given to `--test-dir`.

The default is `tests/phpunit` and files matches, but with an absolute
paths, all files are filtered out, hence the difference in behavior.

Uses relative path for `--test-dir`.

Follows-up to 01565b366f8d

Bug: T288396
Bug: T402398
Change-Id: Ie39c346a1fc93edaf7e089b6a9dfd9b2d80bf51c