tox: lint with Python3, Zuul with Python 2.7
The flake8 linting is still using Python 2.7 because Zuul itself still
run Python 2.7.

The scripts in ./utils that do have a `.py` extension are not discovered
by flake8 and are not linted. They are otherwise not Python 2.7
compatible nor should they.

This change creates a new `lint-zuul` environment which checks the two
files that must remain python2.7:

1) zuul/parameter_functions.py : run by the Zuul daemon
2) tests/test_zuul_scheduler.py : integration tests with Zuul scheduler

They are passed to flake8 directly.

The `lint` environment switches to python3. Since the two files above do
not work under python 3, exclude them using `--extend-exclude`.

The comma separated list is added as a config that can be referred to
from both places. That saves us from having to repeat the files.

* Add the new `lint-zuul` to the list of environments to run.
* Update tests/test_zuul_layout.py `except Exception, e` syntax which is
  no more valid in Python 3.
* Upgrade flake8 to 4.0.1 in order to support recentish Python 3.

Change-Id: I4bc03bccac687900344312e2b2e9c542b198418f