tox: detect and lint python scripts without .py
The utils/docker-updates script is written in python. It is not linted
cause flake8 discovery only check whether a filename matches '*.py'.

We can find those scripts by looking at the shebang:

  $ git grep -l -P '^#!.*python(3|$)'
  dockerfiles/catalyst/deploy_env.py
  dockerfiles/quibble-bullseye-php81-coverage/clover-edit.py
  dockerfiles/quibble-bullseye-php81-coverage/phpunit-junit-edit.py
  dockerfiles/quibble-bullseye-php81-coverage/phpunit-suite-edit.py
  dockerfiles/tox-buster/cobertura-clover-transform.py
  utils/archive-repo.py
  utils/docker-updates
  utils/extract-shell-scripts.py
  utils/zuul-mw-jobs-runner.py
  zuul-migrate

This yields .py files which are already discovered by flake8. `git grep`
can be passed a pathspec (see gitglossary(7)) which excludes files based
on a pattern. The magic (literally) pattern is: :!*.py

I am plumbing those togethers to form a command that runs flake8 against
python scripts that are not having `.py` as a suffix.

Output:

  $ tox -e lint
  ...
  lint: commands[2]> flake8 --extend-exclude \
                     zuul/parameter_functions.py,./tests/test_zuul_scheduler.py
  lint: commands[3]> /bin/bash -c \
    'flake8 $(git grep -l -P '"'"'^#!.*python(3|$)'"'"' '"'"':!*.py'"'"')'

    lint: OK (1.28=setup[0.03]+cmd[0.17,0.27,0.32,0.49] seconds)
    congratulations :) (1.33 seconds)

That is a lot of quoting =]

Change-Id: I6b355e693a1c8a45f3721db8414a837c3049dffb