jjb: use templates for ext/skins coverage publish
The mwext-phpunit-coverage-publish and mwskin-phpunit-coverage-publish
are closely similar. Convert them to a single job template with
parameters which makes it easier to spot difference.

Inline the publisher macros in the template, that makes it easier to
follow. Set the path with a single variable `cover_dest`.

This is a noop in jjb. No jobs are affected.

Bug: T402398
Change-Id: I3fcd6f14474d61ded059aa5aa8f1f27976d99927