jenkins-jobs: default to use `python3` to invoke tox
The script was made to invoke `python3.11` whenever it is present but it
does not necessarily have tox installed.  On CI tox-v4:0.3 image tox is
installed with the default python which is 3.9.

This change the logic to only use python3.11 when python3 is 3.12 or
later. I have used a oneliner given `sys.version_info` can be used for
comparison with a tuple, that is quite handy.

I have dropped `python`, I am assuming all installs have `python3` which
let us remove the Homebrew specific logic and simplify the script.

Follows-up to: baf83540dbc49d81d884cf1d56fcc2fb899bd91b

Change-Id: Id8cf9d2c136194128f50d8ea304732ccbc0fb8df