jjb: move skins test coverage to /cover-extensions/
In order to simplify the CI stack, this move the PHPUnit test coverage
reports of skins along the extensions ones.

On the documentation host, we have symbolic links in place: skins can
already be seen from https://doc.wikimedia.org/cover-extensions/.

This change the CI job to publish directly to /cover-extension/, we can
then swap the symbolic links the other way around. Later, we can change
skins to use the same job as extensions.

Affected jobs:
 - mwskin-phpunit-coverage-publish

Skins with coverage:
- CologneBlue
- MinervaNeue
- Modern (broken)
- MonoBook
- Nostalgia
- Timeless
- Vector
- WikimediaApiPortal

Bug: T402398
Change-Id: Ic3a95c762be1482cd49eabd8cc25b3805dbb9eae