jjb: switch skins coverage to mwext-phpunit-coverage
This is an intermediate step to move the skins coverage report to use
the same system as extensions.

The entrypoint for extensions coverage was recently changed from the
legacy phpunit.php to vendor/bin/phpunit which this change implies.
See I835e5cac8557ccb0e93ea860aee5deb759f48551

The mwext-phpunit-coverage script invokes either:
* the `extensions` testsuite which already runs any tests and parser
  tests.
* in code health context, it uses `extensions:unit` which is made to
  recognizes skins with Ifaeb8c2e278acca89a92187ac39a6c0e1eb503f9

This will later on let us switch to the same job for both extensions and
skins and remove this last skin specific job.

Affected jobs:
 - mwskin-phpunit-coverage-publish

Bug: T395470
Bug: T402398
Depends-On: Ifaeb8c2e278acca89a92187ac39a6c0e1eb503f9
Change-Id: Ifb9eb945cd6f55b7edca01d78bd3d17099ed7e4f