Zuul: [mediawiki/extensions/CampaignEvents] Add CentralAuth dependency
In I8e448612c58b77b7f I left it out saying that «The extension is
already tested with CentralAuth as part of the gate», but that was
probably just my subconscious expressing what I really wish were true.
CentralAuth isn't gated and it doesn't look like T333541 will be
resolved anytime soon. T406582 has a trimmed down proposal, but that's
core-only anyway. In this situation, testing the extension
functionalities with CentralAuth is more important, as that's how the
extension works in production.

So, explicitly add CentralAuth as a dependency for CampaignEvents. Note
that CentralAuth is already installed for CampaignEvents patches as a
transitive dependency, but that is going to be dropped as part of
T389998, hence the explicit addition here.

Bug: T404995
Bug: T389998
Change-Id: Ia2471069b845e1416739fdf1c34b235876b787ed