Add camilojdiaz to CI allowlist
I added a "New developers" section to the allowlist. None of the
the existing sections apply to new contributors who aren't
employees or SDoC/Outreachy participants; on the one hand, many
new developers don't stay for long and we probably don't want the
"trusted users" section to be full of people with 1-2 commits of
whom we actually don't know much about. On the other hand, it's
new developers who are the most negatively affected by CI not
working for them, and where attrition is the highest, and it would
be a shame to not be able to allowlist new developers who make good
contributions. So maybe a dedicated section that can occasionally be
reviewed and cleaned is a good compromise.

Change-Id: I6ced08bd2d079fd15397a112acb3adb662c27334