Zuul: add option to opt out recursive dependencies
As found by Timo in Idf670d490543352de1b18b1a1647e73a6e1e1516, we can
not just turn off recursive processing of dependencies.  Since CI runs
any test it can find, tests of a dependency might end up failing because
its dependencies are not present.

Via T389998 we worked on making tests to be skipped when a required
extension is not loaded and made some Resource Loader definitions to be
optionally registered.  In some case, explicitl dependencies were added
to the mapping.

Although for most extensions we can now disable the recursive
processing, there is a long tail of extensions that needs further work.
Out of 314 extensions defined in zuul/dependencies.yaml, there are still
26 of them that do not pass without recursion.

To smooth the transition, this change makes it possible to disable the
recursive processing on a per extension basis.  By opting in extensions
that are known to pass, we prevent regressions and will more easily
track the work that remains to be done.  This is done by adding another
level in the mapping with key 'recurse' and 'dependencies', so when
previously we had:

  ArticlePlaceholder:
   - Scribunto
   - Wikibase

The mapping becomes:

  ArticlePlaceholder:
    recurse: true
    dependencies:
    - Scribunto
    - Wikibase

Assuming tests pass with just ArticlePlaceholder, Scribunto and
Wikibase, we can then opt out by setting `recurse` to `false`.

Bug: T389998
Change-Id: I15b3d6ea0991dc21926df749db89fb380e1df370